save route database function
