+++ robosistance/src/db/models.rs
@@ -1,5 +1,11 @@
 use mongodb::bson::Uuid;
 use serde::{Deserialize, Serialize};
+use std::fmt;
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct Route {
+    pub commands: Vec<Command>,
+}
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct RobotDetails {
@@ -37,4 +43,28 @@ pub enum MovementData {
     DistanceMoved(u64), //Assuming millimetre granularity.
     RotationAngle(f32),
     PatrolStatus,
+}
+
+#[derive(Copy, Clone, Debug, Serialize, Deserialize)]
+pub enum Command {
+    Hello,
+    Rotate(f32),
+    Move(u64),
+    Beep,
+    Patrol(usize), // As an id.
+    Closed,
+}
+
+impl fmt::Display for Command {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        use Command::*;
+        match self {
+            Hello => write!(f, "Hello"),
+            Rotate(_) => write!(f, "Rotate"),
+            Move(_) => write!(f, "Move"),
+            Beep => write!(f, "Beep"),
+            Patrol(_) => write!(f, "Patrol"),
+            Closed => write!(f, "Closed"),
+        }
+    }
 }
